Reorganize bandcamp files
