+++ src/bandcamp/modules/main.js
@@ -1,8 +1,8 @@
-import { logStorage } from "../modules/storage.js";
-import { injectCSSFile, injectJSFile } from "../modules/utils.js";
+import { logStorage } from "../../modules/storage.js";
+import { injectCSSFile, injectJSFile } from "../../modules/utils.js";
 import { PageTypeDetector } from "./bandcamp.js";
-import { setupPageAlbum } from "./page-album.js";
-import { setupPageMusic } from "./page-music.js";
+import { setupPageAlbum } from "../pages/page-album.js";
+import { setupPageMusic } from "../pages/page-music.js";
 
 export function main () {
   logStorage();
