+++ src/discogs/discogs-csv.js
@@ -1,4 +1,4 @@
-import { keywordsToDiscogsGenres, keywordsToDiscogsStyles } from "../bandcamp/bandcamp.js";
+import { keywordsToDiscogsGenres, keywordsToDiscogsStyles } from "../bandcamp/modules/bandcamp.js";
 import { capitalizeEachWord } from "../modules/utils.js";
 import { Release, Track } from "../app/release.js";
