+++ src/modules/storage.js
@@ -1,5 +1,5 @@
 import { Release } from "../app/release.js";
-import { isValidBandcampURL } from "../bandcamp/html.js";
+import { isValidBandcampURL } from "../bandcamp/modules/html.js";
 import { generateKeyForRelease, generateKeyForUrl, generateKeyUrlMapFromUrls, generateKeysFromUrls } from "./key-generator.js";
 import { isArray, isFunction, isObject } from "./utils.js";
